Devlp (#9)
Доработано

* Ogurtsov 6 var 1_lab
