+++ readme.md
@@ -6,6 +6,7 @@
  - `git log --pretty=format:'%h %ad | %s%d [%an]' --graph --date=short`
  - `git log --graph --all --format=format:'%C(bold blue)%h%C(reset) — %C(bold green)(%ar)%C(reset) %C(white)%s%C(reset) %C(bold white)— %an%C(reset)%C(bold yellow)%d%C(reset)' --abbrev-commit --date=relative`
  - [еще немного полезных команд](https://tproger.ru/translations/most-common-git-screwupsquestions-and-solutions/amp/)
+ - [ПРОЧИТАТЬ](https://habr.com/post/125999/)
  
 ## Как начать работать
